Fixed the assignment button being enabled when the assignment window is already open. Fixed the delete assignment breaking the program. Fixed a lot of nullpointerexceptions when moving from table to table. Fixed the the assignments not getting the correct percent in the gradebook window. Fixed when adding a new gradebook the text field was filled with a 2. Moved un used class files to oldfiles folder. Generated a test .jar file to see if any issues still exist.
Last thing that I would like to change would to be able to set the student score to null so that when adding a new student they don't have 100% on every assignment but can still have a A+ starting out. It also would fix when adding a new assignment everybody getting a 100% by default.